+++ src/payload/mod.rs
@@ -1,3 +1,4 @@
+mod bf16;
 mod differentiable;
 mod elementary;
 mod gemm;
@@ -7,6 +8,7 @@ mod storage;
 mod tensor;
 mod tensorial;
 
+pub use bf16::Bf16;
 pub use differentiable::Differentiable;
 pub use elementary::{Elementary, MapOperation};
 pub use gemm::GemmTask;